social/nostr/runecast/requirements.txt
rich
tornado
pynostr
python-dotenv

